+++ compatibility/step_definitions/multiple-features.ts
@@ -1,0 +1,5 @@
+import { Given } from "@badeball/cypress-cucumber-preprocessor";
+
+Given("an order for {string}", () => {
+  // no-op
+});