+++ lib/add-cucumber-preprocessor-plugin.ts
@@ -5,14 +5,18 @@ import { generateMessages } from "@cucumber/gherkin";
 import { IdGenerator, SourceMediaType } from "@cucumber/messages";
 import parse from "@cucumber/tag-expressions";
 import { getSpecs } from "find-cypress-specs";
+import { v4 as uuid } from "uuid";
 
 import { INTERNAL_PROPERTY_NAME, INTERNAL_SUITE_PROPERTIES } from "./constants";
 import {
   TASK_CREATE_STRING_ATTACHMENT,
   TASK_FRONTEND_TRACKING_ERROR,
   TASK_SPEC_ENVELOPES,
+  TASK_SUGGESTION,
   TASK_TEST_CASE_FINISHED,
   TASK_TEST_CASE_STARTED,
+  TASK_TEST_RUN_HOOK_FINISHED,
+  TASK_TEST_RUN_HOOK_STARTED,
   TASK_TEST_STEP_FINISHED,
   TASK_TEST_STEP_STARTED,
 } from "./cypress-task-definitions";
@@ -31,8 +35,11 @@ import {
   frontendTrackingErrorHandler,
   OnAfterStep,
   specEnvelopesHandler,
+  suggestion,
   testCaseFinishedHandler,
   testCaseStartedHandler,
+  testRunHookFinishedHandler,
+  testRunHookStartedHandler,
   testStepFinishedHandler,
   testStepStartedHandler,
 } from "./plugin-event-handlers";
@@ -114,6 +121,11 @@ export async function addCucumberPreprocessorPlugin(
       config,
       options,
     ),
+    [TASK_TEST_RUN_HOOK_STARTED]: testRunHookStartedHandler.bind(null, config),
+    [TASK_TEST_RUN_HOOK_FINISHED]: testRunHookFinishedHandler.bind(
+      null,
+      config,
+    ),
     [TASK_TEST_CASE_FINISHED]: testCaseFinishedHandler.bind(null, config),
     [TASK_CREATE_STRING_ATTACHMENT]: createStringAttachmentHandler.bind(
       null,
@@ -123,6 +135,7 @@ export async function addCucumberPreprocessorPlugin(
       null,
       config,
     ),
+    [TASK_SUGGESTION]: suggestion.bind(null, config),
   });
 
   const tags = getTags(config.env);
@@ -188,6 +201,8 @@ export async function addCucumberPreprocessorPlugin(
   if (preprocessor.dryRun) {
     config.supportFile = false;
   }
+
+  config.env["testRunStartedId"] = uuid();
 
   return config;
 }