+++ lib/cypress-task-definitions.ts
@@ -27,6 +27,16 @@ export const TASK_TEST_STEP_FINISHED =
 
 export type ITaskTestStepFinished = messages.TestStepFinished;
 
+export const TASK_TEST_RUN_HOOK_STARTED =
+  "cypress-cucumber-preprocessor:test-run-hook-started";
+
+export type ITaskTestRunHookStarted = messages.TestRunHookStarted;
+
+export const TASK_TEST_RUN_HOOK_FINISHED =
+  "cypress-cucumber-preprocessor:test-run-hook-finished";
+
+export type ITaskTestRunHookFinished = messages.TestRunHookFinished;
+
 export const TASK_CREATE_STRING_ATTACHMENT =
   "cypress-cucumber-preprocessor:create-string-attachment";
 
@@ -34,6 +44,10 @@ export const TASK_FRONTEND_TRACKING_ERROR =
   "cypress-cucumber-preprocessor:frontend-tracking-error";
 
 export type ITaskFrontendTrackingError = string;
+
+export const TASK_SUGGESTION = "cypress-cucumber-preprocessor:suggestion";
+
+export type ITaskSuggestion = messages.Suggestion;
 
 export interface ITaskCreateStringAttachment {
   data: string;